// frontend/src/app/shared/services/api/models/userSerilizer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Zidnahum Hudaa Project API
 * OpenAPI spec version: 1.0.0
 */

export interface UserSerilizer {
  readonly id: number;
  /**
     * مطلوب. 150 رمزاً أو أقل، مكونة من حروف وأرقام و @/./+/-/_ فقط
     * @maxLength 150
     * @pattern ^[\w.@+-]+$
     */
  username: string;
  /** @maxLength 150 */
  first_name?: string;
  /** @maxLength 150 */
  last_name?: string;
  /** يحدد ما إذا كان يمكن للمستخدم الدخول إلى موقع الإدارة هذا. */
  is_staff?: boolean;
  /** يقضي بأن هذا المستخدم يمتلك كافة الصلاحيات دون الحاجة لمنحها له تصريحاً. */
  is_superuser?: boolean;
  /** المجموعات التي ينتمي إليها هذا المستخدم. يحصل المستخدم على كافة الصلاحيات الممنوحة لكل مجموعة ينتمي إليها. */
  groups?: number[];
}